Fix theme provider
